+++ python/torch_mlir/extras/fx_importer.py
@@ -354,14 +354,6 @@ def is_builtin_function_or_method(obj: Any) -> bool:
 class InputInfo:
     """Provides additional metadata when resolving inputs."""
 
-    __slots__ = [
-        "program",
-        "input_spec",
-        "node",
-        "ir_type",
-        "mutable_producer_node_name",
-    ]
-
     program: torch.export.ExportedProgram
     input_spec: TypingInputSpec
     node: Node
@@ -915,38 +907,49 @@ class ContextCache:
             tensor_meta = node.meta.get("tensor_meta")
             val = node.meta.get("val")
             sparsity = node.meta.get("sparsity", None)
-            if tensor_meta is not None:
-                assert isinstance(tensor_meta, TensorMetadata)
-                # Quantized tensor meta data is not preserved in our lowering,
-                # so throw error instead of silently doing wrong thing.
-                if tensor_meta.is_quantized:
-                    raise NotImplementedError(
-                        f"Quantized tensor meta data is not supported."
-                    )
-                else:
-                    return self.tensor_metadata_to_type(
-                        tensor_meta, sparsity=sparsity, mutable=mutable
-                    )
-            elif val is not None:
-                # some nodes with symbolic inputs pass a 'val' attribute rather than
-                # tensor_meta
-                if isinstance(val, TorchFakeTensor):
-                    return self.get_vtensor_type(
-                        val.size(), val.dtype, sparsity=sparsity, mutable=mutable
-                    )
-
-                t = SCALAR_TYPE_TO_TORCH_MLIR_TYPE.get(type(val))
-                if t is not None:
-                    return IrType.parse(t, self._c)
-
-            raise NotImplementedError(
-                f"FIXME: Unsupported placeholder node (this often indicates that a necessary) "
-                f"fx preprocessing pass was not run): {node.meta}"
+            return self.value_info_to_type(
+                val, tensor_meta=tensor_meta, sparsity=sparsity, mutable=mutable
             )
         except KeyError as e:
             raise RuntimeError(
                 f"FIXME: Illegal access to torch.fx.Node.meta: {e} ({node.meta.keys()} : {node.meta})"
             )
+
+    def value_info_to_type(
+        self,
+        val,
+        *,
+        tensor_meta: Optional[TensorMetadata] = None,
+        sparsity=None,
+        mutable: bool = False,
+    ):
+        if tensor_meta is not None:
+            assert isinstance(tensor_meta, TensorMetadata)
+            # Quantized tensor meta data is not preserved in our lowering,
+            # so throw error instead of silently doing wrong thing.
+            if tensor_meta.is_quantized:
+                raise NotImplementedError(
+                    f"Quantized tensor meta data is not supported."
+                )
+            else:
+                return self.tensor_metadata_to_type(
+                    tensor_meta, sparsity=sparsity, mutable=mutable
+                )
+        elif val is not None:
+            # some nodes with symbolic inputs pass a 'val' attribute rather than
+            # tensor_meta
+            if isinstance(val, TorchFakeTensor):
+                return self.get_vtensor_type(
+                    val.size(), val.dtype, sparsity=sparsity, mutable=mutable
+                )
+        else:
+            t = SCALAR_TYPE_TO_TORCH_MLIR_TYPE.get(type(val))
+            if t is not None:
+                return IrType.parse(t, self._c)
+        raise NotImplementedError(
+            f"Could not deduce type from value info: "
+            f"tensor_meta={tensor_meta}, val={val}, sparsity={sparsity}"
+        )
 
     def tensor_metadata_to_type(
         self,
@@ -1624,10 +1627,9 @@ class GraphNodeImporter:
             # short-circuit above. Note that if we ever choose to also fully reify Python
             # level result tuples, we will need to create a tuple-boxed version of this and
             # redirect to it for generic object access.
-
             result_types = []
             for v in node.meta["val"]:
-                result_types.append(self._cc.tensor_metadata_to_type(v))
+                result_types.append(self._cc.value_info_to_type(v))
             result_types = tuple(result_types)
         return result_types
